Create AiCWebserv.py
This is a simple webserver for the WiPy, aimed at control applications. The accompanian webpage consist of two input elements, one of type range and one of type number. Changing either of them will generate an Ajax get to the server who responds with the value of the input element. That value is then imposed to the other element. The necessary javascript is in the header section of the webpage.